change FeatureLoc to have NULL phase
git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/pathsoft/artemis/trunk@4687 ee4ac58c-ac51-4696-9907-e4b3aa274f04

+++ uk/ac/sanger/artemis/chado/ChadoDemo.java
@@ -340,7 +340,7 @@ public class ChadoDemo
         int fmin = loc.getFmin() + 1;
         int fmax = loc.getFmax();
         rowData[i][4] = fmin + "..." + fmax;
-        rowData[i][5] = Integer.toString(loc.getStrand());
+        rowData[i][5] = Integer.toString(loc.getStrand().shortValue());
       }
       
       String schema = feature.getOrganism().getAbbreviation().toLowerCase();
